♻️ Refactor DFA and indomitable submit to submissions API
