+++ src/routes/ngs-api/duels/indomitable/[category=categoryIndomitable]/+server.ts
@@ -1,22 +1,11 @@
 import type { IndomitableRun } from '$lib/types/api/duels/indomitable';
-import { parseWeapon } from '$lib/types/api/weapon';
 import { leaderboardDb } from '$lib/server/db/db';
 import { error, json } from '@sveltejs/kit';
 import { convertTimeToRunTime } from '$lib/server/db/util/datetime';
-import { normalizeYoutubeLink } from '$lib/utils/youtube.js';
 import { jsonError } from '$lib/server/error.js';
-import { dbValToWeaponsMap, weaponsToDbValMap } from '$lib/server/db/util/weaponType.js';
-import { notifyDiscordNewRunSubmitted } from '$lib/server/discordNotify.js';
+import { dbValToWeaponsMap } from '$lib/server/db/util/weaponType.js';
 import { dbValToClassMap } from '$lib/server/db/util/ngsClass.js';
 import type { PlayerInfo } from '$lib/types/api/playerInfo.js';
-import {
-	indomitableSubmissionRequestSchema,
-	type IndomitableSubmissionRequest
-} from '$lib/server/types/validation/indomitableSubmissions.js';
-import {
-	checkIndomitableVideoExists,
-	insertIndomitableSoloRun
-} from '$lib/server/repositories/indomitableSubmissionsRepository.js';
 import { RunCategories, parseRunCategory } from '$lib/types/api/categories.js';
 
 const indomitableTables: { [key: string]: string } = {
@@ -199,81 +188,4 @@ const mapData = (queryData: any[]): IndomitableRun[] => {
 			submitter: submitter
 		} as IndomitableRun;
 	});
-};
-
-const indomitableBosses: { [key: string]: string } = {
-	[RunCategories.IndomitableNexAelio]: 'Nex Aelio',
-	[RunCategories.IndomitableRenusRetem]: 'Renus Retem',
-	[RunCategories.IndomitableAmsKvaris]: 'Ams Kvaris',
-	[RunCategories.IndomitableNilsStia]: 'Nils Stia'
-};
-
-/** @type {import('./$types').RequestHandler} */
-export async function POST({ params, request }) {
-	const category = parseRunCategory(params.category);
-	if (!category) {
-		return jsonError(404, 'Unknown category');
-	}
-
-	// Validate request
-	const body = await request.json();
-	let parsedRun: IndomitableSubmissionRequest;
-	try {
-		parsedRun = await indomitableSubmissionRequestSchema.validate(body);
-	} catch (err: any) {
-		return jsonError(400, {
-			error: 'bad_request',
-			details: err.errors
-		});
-	}
-	if (!parsedRun) {
-		return jsonError(400, { error: 'json_parse_error' });
-	}
-
-	// Transform data
-	parsedRun.players.forEach((p) => {
-		p.povVideoLink = p.povVideoLink ? normalizeYoutubeLink(p.povVideoLink) : null;
-		p.weapons = p.weapons.map((w) => weaponsToDbValMap[parseWeapon(w) ?? '']);
-	});
-
-	// Check run data
-	const validationErrors = await checkRunData(parsedRun);
-	if (validationErrors.length > 0) {
-		return jsonError(400, { error: 'bad_request', details: validationErrors });
-	}
-
-	// Insert run
-	try {
-		const pool = await leaderboardDb.connect();
-		const request = pool.request();
-		await insertIndomitableSoloRun(request, parsedRun, category);
-
-		const player1Name = parsedRun.username;
-		const bossName = indomitableBosses[category];
-		notifyDiscordNewRunSubmitted(player1Name, `Indomitable ${bossName} Duel`);
-		return json({ data: 'success' });
-	} catch (err) {
-		console.error(err);
-		throw jsonError(500, { error: 'internal_server_error' });
-	}
-}
-
-const checkRunData = async (run: IndomitableSubmissionRequest) => {
-	const pool = await leaderboardDb.connect();
-	const errorList: string[] = [];
-
-	// Video links not already in use
-	const videoLinks = run.players
-		.map((p) => p.povVideoLink)
-		.filter((l): l is string => l !== undefined);
-
-	const existingVideoLinks = await checkIndomitableVideoExists(pool.request(), videoLinks);
-
-	if (existingVideoLinks.length > 0) {
-		existingVideoLinks.forEach((l) => {
-			errorList.push(`Video already used in another run. VideoUrl=${l}`);
-		});
-	}
-
-	return errorList;
 };
